UIのリファクタリング: サイドバー項目の管理を改善し、ライブラリセクションを整理
