+++ KanadeApp/Sources/Features/Settings/SettingsView.swift
@@ -16,52 +16,8 @@ struct SettingsView: View {
 
     var body: some View {
         @Bindable var appState = appState
-        let discovery = appState.serverDiscovery
 
         Form {
-            if !discovery.servers.isEmpty || discovery.isBrowsing {
-                Section {
-                    if discovery.servers.isEmpty {
-                        HStack {
-                            Spacer()
-                            ProgressView()
-                                .controlSize(.small)
-                            Text("Scanning...")
-                                .foregroundStyle(.secondary)
-                            Spacer()
-                        }
-                    } else {
-                        ForEach(discovery.servers) { server in
-                            Button {
-                                appState.serverAddress = server.host
-                                appState.serverPort = server.port
-                            } label: {
-                                HStack {
-                                    VStack(alignment: .leading, spacing: 2) {
-                                        Text(server.name)
-                                            .font(.headline)
-                                        Text("\(server.host):\(server.port)")
-                                            .font(.caption)
-                                            .foregroundStyle(.secondary)
-                                    }
-                                    Spacer()
-                                    if server.persistent {
-                                        Image(systemName: "pin.fill")
-                                            .foregroundStyle(.orange)
-                                            .font(.caption)
-                                    }
-                                    Circle()
-                                        .fill(Color.green)
-                                        .frame(width: 8, height: 8)
-                                }
-                            }
-                        }
-                    }
-                } header: {
-                    Text("Discovered on Local Network")
-                }
-            }
-
             Section("Server Connection") {
                 LabeledContent("Status") {
                     HStack(spacing: 8) {
@@ -202,7 +158,6 @@ struct SettingsView: View {
                 appState.trustedCAData = data
             }
         }
-        .onAppear { discovery.startBrowsing() }
-        .onDisappear { discovery.stopBrowsing() }
     }
 }
+
